api/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
flask_cors==3.0.10
requests
numpy
pandas
openai